Refactor numeric codecs into SRP submodules
